Bug 709192 - ID 403
The sign extension boted in the writeup is correct, but the specimen
file does no behave as described. Returning to the main loop to input
more data exhausts the file and the interpreter exits. This happens
very quickly.

So another AI hallucination it would seem. This addresses the 'problem'
of the type confusion but obviously since the original file doesn't feil
it makes no difference.

+++ pcl/pxl/pxsessio.c
@@ -1,4 +1,4 @@
-/* Copyright (C) 2001-2023 Artifex Software, Inc.
+/* Copyright (C) 2001-2026 Artifex Software, Inc.
    All Rights Reserved.
 
    This software is provided AS-IS with no warranty, either express or
@@ -756,7 +756,7 @@ pxVendorUnique(px_args_t * par, px_state_t * pxs)
     int code = 0;
 
     if (par->pv[1]) {
-        ulong len = par->pv[1]->value.i;
+        int32 len = par->pv[1]->value.i;
         ulong copy = min(len - par->source.position,
                          par->source.available);
         par->source.data += copy;
